Extract DecodedRequestFrame and DecodedResponseFrame interfaces
Why:
We are going to introduce a kroxylicious-filter-api module that should be
netty independant if possible, this will contain all the interfaces you need
as a Filter Author to build and configure your new filter.
