+++ kroxylicious/src/main/java/io/kroxylicious/proxy/frame/DecodedRequestFrame.java
@@ -9,31 +9,7 @@ import org.apache.kafka.common.message.RequestHeaderData;
 import org.apache.kafka.common.protocol.ApiMessage;
 
 /**
- * A decoded request frame.
+ * docco
  */
-public class DecodedRequestFrame<B extends ApiMessage>
-        extends DecodedFrame<RequestHeaderData, B>
-        implements RequestFrame {
-
-    private final boolean decodeResponse;
-
-    public DecodedRequestFrame(short apiVersion,
-                               int correlationId,
-                               boolean decodeResponse,
-                               RequestHeaderData header,
-                               B body) {
-        super(apiVersion, correlationId, header, body);
-        this.decodeResponse = decodeResponse;
-    }
-
-    @Override
-    public short headerVersion() {
-        return apiKey().messageType.requestHeaderVersion(apiVersion);
-    }
-
-    @Override
-    public boolean decodeResponse() {
-        return decodeResponse;
-    }
-
+public interface DecodedRequestFrame<B extends ApiMessage> extends DecodedFrame<RequestHeaderData, B> {
 }
