+++ kroxylicious/src/main/java/io/kroxylicious/proxy/frame/NettyDecodedResponseFrame.java
@@ -1,0 +1,25 @@
+/*
+ * Copyright Kroxylicious Authors.
+ *
+ * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
+ */
+package io.kroxylicious.proxy.frame;
+
+import org.apache.kafka.common.message.ResponseHeaderData;
+import org.apache.kafka.common.protocol.ApiMessage;
+
+/**
+ * A decoded response frame.
+ */
+public class NettyDecodedResponseFrame<B extends ApiMessage>
+        extends NettyDecodedFrame<ResponseHeaderData, B>
+        implements ResponseFrame, DecodedResponseFrame<B> {
+
+    public NettyDecodedResponseFrame(short apiVersion, int correlationId, ResponseHeaderData header, B body) {
+        super(apiVersion, correlationId, header, body);
+    }
+
+    public short headerVersion() {
+        return apiKey().messageType.responseHeaderVersion(apiVersion);
+    }
+}
