+++ kroxylicious/src/main/java/io/kroxylicious/proxy/internal/filter/FetchResponseTransformationFilter.java
@@ -29,7 +29,6 @@ import org.slf4j.LoggerFactory;
 
 import io.kroxylicious.proxy.filter.FetchResponseFilter;
 import io.kroxylicious.proxy.filter.KrpcFilterContext;
-import io.kroxylicious.proxy.internal.util.NettyMemoryRecords;
 
 /**
  * An filter for modifying the key/value/header/topic of {@link ApiKeys#FETCH} responses.
@@ -110,7 +109,7 @@ public class FetchResponseTransformationFilter implements FetchResponseFilter {
                         TimestampType.CREATE_TIME, 0)) {
 
                     for (MutableRecordBatch batch : records.batches()) {
-                        for (Iterator<Record> batchRecords = batch.iterator(); batchRecords.hasNext(); ) {
+                        for (Iterator<Record> batchRecords = batch.iterator(); batchRecords.hasNext();) {
                             Record batchRecord = batchRecords.next();
                             newRecords.append(batchRecord.timestamp(), batchRecord.key(), valueTransformation.transform(topicData.topic(), batchRecord.value()));
                         }
